Fix qwen_image_edit_2509_manga_tone: cast transformer to target dtype after LoRA fuse to avoid dtype mismatch

+++ qwen_image_edit_2509_manga_tone/pytorch/loader.py
@@ -80,7 +80,7 @@ class ModelLoader(ForgeModel):
         )
         pipe.load_lora_weights(LORA_REPO_ID, weight_name=LORA_WEIGHT_NAME)
         pipe.fuse_lora()
-        self._transformer = pipe.transformer
+        self._transformer = pipe.transformer.to(dtype=dtype)
         self._transformer.eval()
         del pipe
         return self._transformer
